requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.44
asyncmy==0.2.10
PyMySQL==1.1.2

pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0

python-dotenv==1.2.1
python-dateutil==2.9.0.post0

passlib[argon2]==1.7.4
PyJWT==2.10.1
email-validator==2.3.0
python-multipart==0.0.20
httpx==0.28.1
itsdangerous==2.2.0

pytest==9.0.1
pytest-asyncio==1.3.0
aiosqlite==0.21.0
openai==2.0.0
